requirements.txt
Django
djangorestframework
psycopg2
requests
responses
PyJWT
djangorestframework-simplejwt
drf-spectacular
django-cors-headers